Renommage publish/subscribe et réorganisation de certains éléments.
