+++ Cerveau/nodes/can_node.py
@@ -1,7 +1,6 @@
 from Cerveau.interfaces.interface_description.python_parser import Frame
 from Cerveau.interfaces.can_interface.can_handler import CanHandler
-from Cerveau.node_manager.node_manager import NodeBase
-from Cerveau.node_manager.event_handler import observe_event
+from Cerveau.engine_base.node_base import NodeBase, subscribe_to_event
 from Cerveau.events.can_events import EventCanReceive, EventCanSend
 
 from typing import Dict
@@ -20,8 +19,8 @@ class CanNode(NodeBase):
 
 	def on_message(self, frame_type: Frame, data_fields: Dict):
 		ev = EventCanReceive(frame_type, data_fields)
-		self.send_event(ev, state=frame_type.cmd_id)
+		self.publish_event(ev, state=frame_type.cmd_id)
 
-	@observe_event(EventCanSend)
+	@subscribe_to_event(EventCanSend)
 	def send_message(self, ev):
 		self.can_handler.send(ev.frame_type, ev.data_fields)
